git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/cfa@5293 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/interval/IntervalAnalysisElement.java
@@ -253,7 +253,7 @@ public class IntervalAnalysisElement implements AbstractElement
 
     // ... and clone the reference count
     for(String variableName : referenceCounts.keySet())
-      newElement.referenceCounts.put(variableName, getReferenceCount(variableName).intValue());
+      newElement.referenceCounts.put(variableName, getReferenceCount(variableName));
 
     return newElement;
   }
